Remove projectstorm/react-diagrams from codebase
