--- nodemapper/src/NodeMap/components/index.ts
@@ -1,17 +1,0 @@
-export * from "./label/DefaultLabelFactory";
-export * from "./label/DefaultLabelModel";
-export * from "./label/DefaultLabelWidget";
-
-export * from "./link/DefaultLinkFactory";
-export * from "./link/DefaultLinkModel";
-export * from "./link/DefaultLinkWidget";
-export * from "./link/DefaultLinkSegmentWidget";
-export * from "./link/DefaultLinkPointWidget";
-
-export * from "./node/DefaultNodeFactory";
-export * from "./node/DefaultNodeModel";
-export * from "./node/DefaultNodeWidget";
-
-export * from "./port/DefaultPortFactory";
-export * from "./port/DefaultPortLabelWidget";
-export * from "./port/DefaultPortModel";
