--- nodemapper/src/NodeMap/components/label/DefaultLabelFactory.tsx
@@ -1,25 +1,0 @@
-import * as React from "react";
-import { DefaultLabelModel } from "./DefaultLabelModel";
-import { DefaultLabelWidget } from "./DefaultLabelWidget";
-import { AbstractReactFactory } from "@projectstorm/react-canvas-core";
-import { DiagramEngine } from "@projectstorm/react-diagrams-core";
-
-/**
- * @author Dylan Vorster
- */
-export class DefaultLabelFactory extends AbstractReactFactory<
-  DefaultLabelModel,
-  DiagramEngine
-> {
-  constructor() {
-    super("default");
-  }
-
-  generateReactWidget(event): JSX.Element {
-    return <DefaultLabelWidget model={event.model} />;
-  }
-
-  generateModel(event): DefaultLabelModel {
-    return new DefaultLabelModel();
-  }
-}
