+++ nodemapper/src/NodeMap/index.ts
@@ -1,2 +1,1 @@
-export * from "./components";
 export * from "./scene";
