+++ nodemapper/src/NodeMap/scene/index.ts
@@ -1,2 +1,1 @@
-export * from "./NodeScene";
 export * from "./NodeMapEngine";
