--- nodemapper/src/gui/Runner/Buttons.css
@@ -1,33 +1,0 @@
-.btn {
-  border: 2px solid #04aa6d;
-  background-color: #dddddd; /* Green */
-  color: #04aa6d;
-  text-align: center;
-  text-decoration: none;
-  vertical-align: middle;
-  display: inline-block;
-  font-size: 12px;
-  margin-left: 6px;
-  margin-right: 6px;
-  padding-left: 4px;
-  padding-right: 4px;
-  border-radius: 4px;
-  min-width: 50px;
-  min-height: 25px;
-  transition-duration: 0.4s;
-}
-
-.btn:hover {
-  background-color: #04aa6d;
-  color: #dddddd;
-}
-
-.btn:active {
-  background-color: #008554;
-  color: #dddddd;
-}
-
-.btn:disabled {
-  background-color: #cccccc;
-  color: #666666;
-}
