--- nodemapper/src/gui/Runner/FileBrowser.css
@@ -1,4 +1,0 @@
-.filebrowser {
-  padding: 2px;
-  padding-top: 20px;
-}
