--- nodemapper/src/gui/Runner/FileBrowser.tsx
@@ -1,38 +1,0 @@
-import React from "react";
-import FolderContents from "./FolderContents";
-
-import { useAppDispatch } from "redux/store/hooks";
-import { runnerLoadWorkflow } from "redux/actions";
-
-import "./FileBrowser.css";
-
-const FileBrowser = () => {
-  const dispatch = useAppDispatch();
-  const btnLoadWorkflow = () => {
-    dispatch(runnerLoadWorkflow());
-  };
-
-  return (
-    <div className="filebrowser">
-      <table style={{ width: "100%" }}>
-        <thead>
-          <tr>
-            <th>
-              <div style={{ textAlign: "left" }}>File Browser</div>
-            </th>
-            <th>
-              <div style={{ textAlign: "right" }}>
-                <button className="btn" onClick={btnLoadWorkflow}>
-                  LOAD
-                </button>
-              </div>
-            </th>
-          </tr>
-        </thead>
-      </table>
-      <FolderContents />
-    </div>
-  );
-};
-
-export default FileBrowser;
