--- nodemapper/src/gui/Runner/FolderContents.css
@@ -1,17 +1,0 @@
-.button {
-  border: 1px;
-  padding: 0;
-  width: 100%;
-  height: 25px;
-  text-align: left;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 12px;
-  background: #555555;
-  color: #cccccc;
-}
-
-.button:hover {
-  background: #777777;
-  color: #cccccc;
-}
