--- nodemapper/src/gui/Runner/Footer.css
@@ -1,7 +1,0 @@
-.footer {
-  padding: 5px;
-  color: #cccccc;
-  background: #333333;
-  width: 100%;
-  height: 100%;
-}
