--- nodemapper/src/gui/Runner/Header.css
@@ -1,19 +1,0 @@
-@import "Buttons.css";
-
-* {
-  margin: 0;
-  padding: 0;
-}
-
-html,
-body,
-#header-panel,
-#header-panel > div {
-  background: #dddddd;
-  font-family: Oswald, sans-serif;
-}
-
-#header-panel,
-#header-panel > div {
-  margin-left: 20px;
-}
