--- nodemapper/src/gui/Runner/MainBody.css
@@ -1,14 +1,0 @@
-.split {
-  width: 100%;
-  position: fixed;
-  z-index: 1;
-  overflow-x: hidden;
-}
-
-.top {
-}
-
-.bottom {
-  top: 80%;
-  height: 20%;
-}
