--- nodemapper/src/gui/Runner/MainPanel.css
@@ -1,16 +1,0 @@
-.split {
-  height: 100%;
-  position: fixed;
-  z-index: 1;
-  overflow-x: hidden;
-}
-
-.left {
-  left: 0;
-  width: 33%;
-}
-
-.right {
-  right: 0;
-  width: 67%;
-}
