--- nodemapper/src/gui/Runner/NodeManager.css
@@ -1,43 +1,0 @@
-. {
-  cursor: pointer;
-}
-
-.diagram-container {
-  background: #333333;
-  width: 100%;
-  height: 100%;
-}
-
-.custom-node {
-  border: solid 2px gray;
-  border-radius: 5px;
-  width: 50px;
-  height: 100px;
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  position: relative;
-}
-
-.custom-node-color {
-  position: absolute;
-  top: 15%;
-  left: 50%;
-  width: 20px;
-  height: 20px;
-  transform: translate(-50%, -50%);
-  border-radius: 10px;
-}
-
-.circle-port {
-  width: 12px;
-  height: 12px;
-  margin: 2px;
-  border-radius: 4px;
-  background: darkgray;
-  cursor: pointer;
-}
-
-.circle-port:hover {
-  background: mediumpurple;
-}
