+++ nodemapper/src/gui/Runner/Runner.tsx
@@ -1,20 +1,8 @@
 import React from "react";
-import Header from "./Header";
-import MainPanel from "./MainPanel";
 
 // Layout for main window, including sliding-pane support
 const Runner = () => {
-  return (
-    <>
-      <div className="header">
-        <Header />
-      </div>
-
-      <div className="main">
-        <MainPanel />
-      </div>
-    </>
-  );
+  return <>Runner</>;
 };
 
 export default Runner;
