--- nodemapper/src/gui/Runner/SidePane.css
@@ -1,18 +1,0 @@
-.sidepane {
-  padding: 2px;
-  padding-top: 0;
-}
-
-.title {
-  font-size: 22px;
-  font-weight: normal;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  height: 40px;
-}
-
-.subtitle {
-  font-size: 18px;
-  margin-top: 2px;
-}
