--- nodemapper/src/gui/Runner/SidePaneContent.css
@@ -1,4 +1,0 @@
-@import "Buttons.css";
-
-body {
-}
