+++ nodemapper/src/redux/actions/display.ts
@@ -1,6 +1,5 @@
 import { createAction } from "@reduxjs/toolkit";
 
-export const displayZoomToFit = createAction("display/zoom-to-fit");
 export const displayDeleteResults = createAction("display/delete-results");
 export const displayGetFolderInfo = createAction("display/get-folder-info");
 export const displayUpdateNodeInfo = createAction<string>(
