+++ nodemapper/src/redux/middleware/runner.ts
@@ -1,17 +1,9 @@
-import RunnerEngine from "gui/Runner/RunnerEngine";
 import * as globals from "redux/globals";
 
 import { displayGetFolderInfo } from "redux/actions";
 import { displayUpdateNodeInfo } from "redux/actions";
 import { displayStoreFolderInfo } from "redux/actions";
 
-import { runnerStoreMap } from "redux/actions";
-import { runnerStoreLint } from "redux/actions";
-import { runnerNodeSelected } from "redux/actions";
-import { runnerLintSnakefile } from "redux/actions";
-import { runnerNodeDeselected } from "redux/actions";
-import { runnerQueryJobStatus } from "redux/actions";
-import { runnerStoreJobStatus } from "redux/actions";
 import { runnerUpdateStatusText } from "redux/actions";
 
 type Query = Record<string, unknown>;
@@ -30,36 +22,6 @@ export const runnerMiddleware = ({ getState, dispatch }) => {
         console.log("Middleware [runner]: ", action);
       }
       switch (action.type) {
-        case "runner/node-selected":
-          NodeSelected(action, dispatch, getState().display.graph_is_moveable);
-          break;
-        case "runner/node-deselected":
-          NodeDeselected(dispatch);
-          break;
-        case "runner/select-none":
-          SelectNone();
-          break;
-        case "runner/import-snakefile":
-          ImportSnakefile(dispatch);
-          break;
-        case "runner/load-snakefile":
-          LoadSnakefile(action, dispatch);
-          break;
-        case "runner/launch-snakefile":
-          LaunchSnakefile(dispatch, getState().display.folderinfo);
-          break;
-        case "runner/query-job-status":
-          QueryJobStatus(dispatch, getState().display.folderinfo);
-          break;
-        case "runner/build-snakefile":
-          BuildSnakefile(dispatch, getState().runner.serialize);
-          break;
-        case "runner/lint-snakefile":
-          LintSnakefile(dispatch, getState().runner.serialize);
-          break;
-        case "runner/load-workflow":
-          LoadWorkflow(dispatch, getState().display.folderinfo);
-          break;
         case "runner/delete-results":
           DeleteResults(dispatch, getState().display.folderinfo);
           break;
@@ -74,224 +36,6 @@ export const runnerMiddleware = ({ getState, dispatch }) => {
 ///////////////////////////////////////////////////////////////////////////////
 // Middleware
 ///////////////////////////////////////////////////////////////////////////////
-
-const NodeSelected = (action, dispatch, graph_is_moveable: boolean): void => {
-  if (!graph_is_moveable) {
-    const runner = RunnerEngine.Instance;
-    const node = runner.getNodeById(action.payload.id);
-    let payload = {};
-    if (node === null) {
-      console.debug("Selected node not found in engine: ", action.payload.id);
-      payload = {
-        id: action.payload.id,
-        name: "ERROR: Failed to find node (" + action.payload.id + ")",
-        type: "ERROR: Failed to find node (" + action.payload.id + ")",
-        code: "ERROR: Failed to find node (" + action.payload.id + ")",
-      };
-    } else {
-      const json = JSON.parse(node.getOptions().extras);
-      payload = {
-        id: action.payload.id,
-        name: node.getOptions()["name"],
-        type: json.type,
-        code: json.content,
-      };
-    }
-    dispatch(displayUpdateNodeInfo(JSON.stringify(payload)));
-  }
-};
-
-const NodeDeselected = (dispatch): void => {
-  dispatch(displayUpdateNodeInfo(""));
-};
-
-const SelectNone = () => {
-  // Link to singleton instance of runner graph engine
-  const runner = RunnerEngine.Instance;
-  runner.NodesSelectNone();
-};
-
-const ImportSnakefile = (dispatch): void => {
-  QueryAndLoadTextFile(async (content, filename) => {
-    const query: Query = {
-      query: "runner/tokenize",
-      data: {
-        format: "Snakefile",
-        content: content,
-      },
-    };
-    const callback = (content: Query) => {
-      RebuildNodeMap(content, dispatch);
-    };
-    switch (backend) {
-      case "rest":
-        SubmitQuery(query, dispatch, callback);
-        break;
-      case "electron":
-        callback((await runnerAPI.TokenizeLoad(query)) as Query);
-        break;
-      default:
-        console.error("Unknown backend: ", backend);
-    }
-    dispatch(displayUpdateNodeInfo(""));
-  });
-};
-
-const LoadSnakefile = async (action, dispatch): Promise<void> => {
-  dispatch(runnerUpdateStatusText("Loading Snakefile..."));
-  const query: Query = {
-    query: "runner/tokenize_load",
-    data: {
-      format: "Snakefile",
-      content: action.payload,
-    },
-  };
-  const callback = (content: Query) => {
-    RebuildNodeMap(content, dispatch);
-  };
-  switch (backend) {
-    case "rest":
-      SubmitQuery(query, dispatch, callback);
-      break;
-    case "electron":
-      callback((await runnerAPI.TokenizeLoad(query)) as Query);
-      break;
-    default:
-      console.error("Unknown backend: ", backend);
-  }
-  dispatch(runnerUpdateStatusText(""));
-};
-
-const LaunchSnakefile = async (dispatch, folderinfo: string): Promise<void> => {
-  const query: Query = {
-    query: "runner/launch",
-    data: {
-      format: "Snakefile",
-      content: JSON.parse(folderinfo).foldername,
-    },
-  };
-  const callback = (content: Query) => {
-    console.info("Launch response: ", content["body"]);
-  };
-  switch (backend) {
-    case "rest":
-      SubmitQuery(query, dispatch, callback);
-      break;
-    case "electron":
-      callback((await runnerAPI.Launch(query)) as Query);
-      break;
-    default:
-      console.error("Unknown backend: ", backend);
-  }
-  setTimeout(() => dispatch(runnerQueryJobStatus()), 5000);
-};
-
-const QueryJobStatus = async (dispatch, folderinfo: string): Promise<void> => {
-  const query: Query = {
-    query: "runner/jobstatus",
-    data: {
-      format: "Snakefile",
-      content: JSON.parse(folderinfo).foldername,
-    },
-  };
-  const callback = (content: Query) => {
-    dispatch(runnerStoreJobStatus(content["body"] as string));
-  };
-  switch (backend) {
-    case "rest":
-      SubmitQuery(query, dispatch, callback);
-      break;
-    case "electron":
-      callback((await runnerAPI.JobStatus(query)) as Query);
-      break;
-    default:
-      console.error("Unknown backend: ", backend);
-  }
-  setTimeout(() => dispatch(runnerQueryJobStatus()), 1000);
-};
-
-const BuildSnakefile = async (dispatch, serial: string): Promise<void> => {
-  const query: Query = {
-    query: "runner/build",
-    data: {
-      format: "Snakefile",
-      content: serial,
-    },
-  };
-  const callback = (content: Query) => {
-    // Download returned content as file
-    const filename = "Snakefile";
-    const element = document.createElement("a");
-    element.setAttribute(
-      "href",
-      "data:text/plain;charset=utf-8," +
-        encodeURIComponent(content["body"] as string)
-    );
-    element.setAttribute("download", filename);
-    element.style.display = "none";
-    document.body.appendChild(element);
-    element.click();
-    document.body.removeChild(element);
-  };
-  switch (backend) {
-    case "rest":
-      SubmitQuery(query, dispatch, callback);
-      break;
-    case "electron":
-      callback((await runnerAPI.Build(query)) as Query);
-      break;
-    default:
-      console.error("Unknown backend: ", backend);
-  }
-};
-
-const LintSnakefile = async (dispatch, serial: string): Promise<void> => {
-  const query: Query = {
-    query: "runner/lint",
-    data: {
-      format: "Snakefile",
-      content: serial,
-    },
-  };
-  const callback = (content: Query) => {
-    dispatch(runnerStoreLint(content["body"] as string));
-  };
-  switch (backend) {
-    case "rest":
-      SubmitQuery(query, dispatch, callback);
-      break;
-    case "electron":
-      callback((await runnerAPI.Lint(query)) as Query);
-      break;
-    default:
-      console.error("Unknown backend: ", backend);
-  }
-};
-
-const LoadWorkflow = async (dispatch, folderinfo: string): Promise<void> => {
-  dispatch(runnerUpdateStatusText("Loading Workflow..."));
-  const query: Query = {
-    query: "runner/loadworkflow",
-    data: {
-      format: "Snakefile",
-      content: JSON.parse(folderinfo).foldername,
-    },
-  };
-  const callback = (content: Query) => {
-    RebuildNodeMap(content, dispatch);
-  };
-  switch (backend) {
-    case "rest":
-      SubmitQuery(query, dispatch, callback);
-      break;
-    case "electron":
-      callback((await runnerAPI.LoadWorkflow(query)) as Query);
-      break;
-    default:
-      console.error("Unknown backend: ", backend);
-  }
-  dispatch(runnerUpdateStatusText(""));
-};
 
 const DeleteResults = async (dispatch, folderinfo: string): Promise<void> => {
   dispatch(runnerUpdateStatusText("Deleting Results..."));
@@ -324,8 +68,10 @@ const DeleteResults = async (dispatch, folderinfo: string): Promise<void> => {
 ///////////////////////////////////////////////////////////////////////////////
 
 const RebuildNodeMap = (content: Query, dispatch): void => {
+  throw new Error("Not implemented");
+
   // Rebuild map from returned (segmented) representation
-  const nodeMapEngine = RunnerEngine.Instance;
+  /*const nodeMapEngine = RunnerEngine.Instance;
   nodeMapEngine.ConstructMapFromBlocks(JSON.parse(content["body"] as string));
   dispatch(runnerStoreMap(content["body"] as string));
   nodeMapEngine.AddSelectionListeners(
@@ -347,7 +93,7 @@ const RebuildNodeMap = (content: Query, dispatch): void => {
     }
   );
   // Submit query to automatically lint file
-  dispatch(runnerLintSnakefile());
+  dispatch(runnerLintSnakefile());*/
 };
 
 const QueryAndLoadTextFile = (
